+++ app/collections/homes/permissions.js
@@ -1,0 +1,10 @@
+import Homes from './';
+
+Homes.allow({
+  insert: function () {
+    return true;
+  },
+  update: function () {
+    return true;
+  }
+});
