+++ app/collections/homes/schema.js
@@ -1,0 +1,10 @@
+import SimpleSchema from 'simpl-schema';
+
+export default HomesSchema = new SimpleSchema({
+  name:{
+    type:String
+  },
+  groupId: {
+    type: String,
+  }
+});
